Migrate to Asp.Net Core Web Api
